Redesign of the layout
