+++ .devcontainer/devcontainer.json
@@ -6,8 +6,8 @@
 	"remoteUser": "node",
 
 	"mounts": [
-		"source=${localWorkspaceFolderBasename}-node_modules,target=${containerWorkspaceFolder}/node_modules,type=volume"
+		"source=${localWorkspaceFolderBasename}-node_modules,target=${containerWorkspaceFolder}/client/node_modules,type=volume"
 	],
 
-	"postCreateCommand": "sudo chown node node_modules && sudo chown node src"
+	"postCreateCommand": "cd client && sudo chown node node_modules && npm install --force"
 }
