+++ client/.env
@@ -1,0 +1,1 @@
+FAST_REFRESH = false
