+++ client/src/Components/Chatbox/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Chatbox = () => {
+  return (
+    <div>Chatbox</div>
+  )
+}
+
+export default Chatbox;
