+++ client/src/Components/Input/InputText/InputText.scss
@@ -1,0 +1,36 @@
+.div-input-text {
+    display: flex;
+    flex-direction: column;
+    position: fixed;
+    bottom: -12px;
+    width: 640px;
+    height: 110px;
+    background-color: var(--grey-main);
+    left: 50%;
+    transform: translate(-50%);
+    border: solid 1px #ffffff35;
+    border-radius: 16px;
+
+    .div-input-text-box {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        margin: 16px 16px 0px 16px;
+        height: 64px;
+        border: solid 1px #ffffff35;
+        border-radius: 8px;
+
+        .div-input-text-box-item1 {
+            margin-left: 16px;
+        }
+        .div-input-text-box-item2 {
+            width: 100%;
+            height: 32px;
+            margin: 0px 8px 0px 8px;
+        }
+        .div-input-text-box-item3 {
+            margin-right: 16px;
+            color: var(--grey-main);
+        }
+    }
+}
