+++ client/src/Components/Props/Icons/Enter/index.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import "./Enter.scss";
+import { AiOutlineEnter } from "react-icons/ai"
+
+const Enter = () => {
+  return (
+    <div className='icon-enter'><AiOutlineEnter /></div>
+  )
+}
+
+export default Enter
