+++ client/src/Components/Props/Icons/Plus/Plus.scss
@@ -1,0 +1,14 @@
+.icon-plus {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 32px;
+    height: 32px;
+    border: solid 1px #ffffff35;
+    border-radius: 40px;
+}
+
+.icon-plus:hover {
+    color: var(--grey-main);
+    background-color: white;
+}
