+++ client/src/Components/Props/Icons/Plus/index.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import "./Plus.scss";
+import { BsPlusLg } from "react-icons/bs"
+
+const Plus = () => {
+  return (
+    <div className='icon-plus'><BsPlusLg /></div>
+  )
+}
+
+export default Plus
