+++ client/src/Routes/Main/Main.scss
@@ -1,0 +1,14 @@
+body{
+    margin: 0;
+    padding: 0;
+    background-color: var(--grey-main);
+}
+.div-main {
+    width: 100%;
+    height: 100vh;
+    background-color: #141414;
+
+    .div-chat-area {
+        
+    }
+}
