+++ client/src/Routes/_404/_404.scss
@@ -1,0 +1,18 @@
+body {
+    color: white;
+}
+
+.div-main-404 {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    margin-top: 25%;
+
+    .div-404 {
+        font-size: xx-large;
+        font-weight: 900;
+    }
+    .div-page-not-found {
+        font-size: large;
+    }
+}
