--- src/Components/Input/InputText/InputText.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-import "./InputText.scss";
-
-const InputText = () => {
-    return (
-        <div class="div-input-text">
-            <input class="input-text-normal" type="text" />
-        </div>
-    )}
-
-export default InputText;
