--- src/Components/Input/InputText/InputText.scss
@@ -1,10 +1,0 @@
-.div-input-text {
-    display: flex;
-    flex-direction: column;
-
-    .input-text-normal {
-        width: 99%;
-        height: 32px;
-        padding: auto;
-    }
-}
