--- src/Routes/Main/Main.scss
@@ -1,202 +1,0 @@
-body{
-    margin: 0;
-    padding: 0;
-}
-.nav{
-    display: grid;
-    grid-template-columns: repeat(100, 1%);
-    grid-auto-rows: 10vh;
-    .nav-left{
-        background-color: #292929;
-        grid-row: 1/11;
-        grid-column: 1/9;
-        display: flex;
-        .div-images{
-            margin-top: 90px;
-            display: flex;
-            width: 100%;
-            flex-direction: column;
-            gap: 2% 0;
-            align-items: center;
-        img{
-            width: 55px;
-            height: 55px;
-            border-radius: 50%;
-        }
-    }
-    }
-.nav-top{
-    background-color: #333333;
-    grid-row: 1/2;
-    grid-column: 9/101;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    border-bottom:2px solid #292929;
-   
-    
-    .nav-top-left{
-        display: flex;
-        align-items: center;
-        justify-content: space-around;
-        width: 13%;
-        i{
-            color: white;
-            font-size: 20px;
-        }
-        img{
-            width: 30px;
-            height: 30px;
-            border-radius: 50%;
-        }
-        span{
-            font-size: 20px;
-            color: white;
-            font-weight: 700;
-            
-        }
-    }
-    .nav-top-middle{
-        border: solid 1px white;
-        width: 25%;
-        margin-right: 5%;
-        height: 60%;
-        display: flex;
-        align-items: center;
-        border-radius: 10px;
-        
-       
-        i{
-            margin-left: 10px;
-            color: white;
-        }
-        input{
-            margin-left: 10px;
-            outline: 0;
-            border: none;
-            background-color: transparent;
-        }
-        input::placeholder{
-            color: white;
-        }
-    }
-    .nav-top-right{
-        width: 2%;
-        margin-right: 10px;
-        i{
-            color: white;
-            font-size: 20px;
-            
-        }
-    }
-    
-}
-.div-main{
-    grid-row: 2/11;
-    grid-column: 9/84;
-    background-color: #333333;
-
-    .text-input{
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        position: absolute;
-        bottom: 10px;
-        width: 75%;
-
-        .text-input-field {
-            width: 98%;
-            height: 32px;
-        }
-
-        .button-message {
-            width: 60px;
-            height: 38px;
-        }
-    }
-}
-.nav-right{
-    grid-row: 2/11;
-    grid-column:84/101;
-    background-color: #292929;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    .div-axton-members{
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        
-        img{
-            height: 80px;
-            width: 80px;
-            border-radius: 50%;
-            margin-top: 10px;
-        }
-        .span-axton{
-            font-size: 19px;
-            color: white;
-            margin-top: 13px;
-        }
-        .span-members{
-            color: white;
-            margin-top: 2px;
-            font-size: 13px;
-            word-spacing: 3px;
-            font-weight: 200;
-
-        }
-        .div-icons{
-            display: flex;
-            justify-content: space-around;
-            gap: 10%;
-            margin-top: 10px;
-            
-            i{
-                color: #333333;
-                padding: 10px;
-                background-color: #e7e9ef;
-                border-radius: 7px;
-            }
-        }
-      
-    }
-    .div-members{
-        height: 60%;
-        width: 90%;
-        margin-top: 15%;
-        background-color: #333333;
-        border-top-left-radius:10px ;
-        border-top-right-radius: 10px;
-        display: flex;
-        flex-direction: column;
-        gap: 1px 0;
-        span{
-            font-size: 15px;
-            margin: 15px 0 0 10px;
-            color: white;
-        }
-        .div-story{
-            display: flex;
-            height: 40px;
-            align-items: center;
-           
-            .span-story{
-             
-                img{
-                    height: 30px;
-                    width: 30px;
-                    padding: 2px;
-                    background-color: #43c079;
-                    border-radius: 50%;
-                }
-            }
-            .span-name{
-                font-size: 12px;
-
-            }
-
-        }
-    }
-}
-}
